Move NameFormElement to domManipulationUtils
We don't use this outside of the client code, and we don't have any
other UI types defined in server/@types/ui.d.ts, so moving this to the
client code (which defines other types) seems sensible for now
