--- server/@types/ui.d.ts
@@ -1,9 +1,0 @@
-type NameFormElementElements = HTMLFormControlsCollection & {
-	name: HTMLInputElement;
-};
-
-type NameFormElement = HTMLFormElement & {
-	elements: NameFormElementElements;
-};
-
-export type { NameFormElement };
